+++ server/portfolios/models.py
@@ -7,7 +7,7 @@ import random
 class Portfolio(models.Model):
     user = models.ForeignKey(User, on_delete=models.CASCADE)
     account_number = models.IntegerField()
-    balance = models.IntegerField()
+    balance = models.DecimalField(max_digits=12, decimal_places=2)
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
